implement query chaining to prevent even the 1 second burst of duplicate questions
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@834 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/lwres.hh
@@ -40,8 +40,8 @@
 #include "dns.hh"
 using namespace std;
 
-int asendto(const char *data, int len, int flags, const ComboAddress& ip, int id, const string& domain, int* fd);
-int arecvfrom(char *data, int len, int flags, const ComboAddress& ip, int *d_len, int id, const string& domain, int fd);
+int asendto(const char *data, int len, int flags, const ComboAddress& ip, uint16_t id, const string& domain, int* fd);
+int arecvfrom(char *data, int len, int flags, const ComboAddress& ip, int *d_len, uint16_t id, const string& domain, int fd);
 
 class LWResException : public AhuException
 {
